feat: create new components that could be reutilized
